# docs/swagger.yaml
definitions:
  errcode.Error:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      err:
        type: error
      msg:
        type: string
    type: object
...
  service.CategoryResponse:
    properties:
      id:
        example: 1
        type: integer
      name:
        example: 电影
        type: string
      workNum:
        example: 1
        type: integer
    type: object
...
  service.CollegeCreateRequest:
    properties:
      introduce:
        type: string
      title:
        type: string
    required:
    - introduce
    - title
    type: object
  service.CollegePageResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/service.CollegeResponse'
        type: array
      total:
        type: integer
    type: object
  service.CollegeResponse:
    properties:
      id:
        type: integer
      introduce:
        type: string
      title:
        type: string
      worksNums:
        type: integer
    type: object
  service.CollegeUpdateRequest:
    properties:
      introduce:
        type: string
      title:
        type: string
    type: object
  service.CollegeWorksPageResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/service.WorkResponse'
        type: array
      total:
        type: integer
    type: object
  service.CommentPageResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/service.CommentResponse'
        type: array
      total:
        type: integer
    type: object
  service.CommentRequest:
    properties:
      content:
        type: string
      objId:
        type: integer
      objType:
        type: integer
    required:
    - content
    - objId
    - objType
    type: object
  service.CommentResponse:
    properties:
      content:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/service.UserResponse'
        type: object
    type: object
  service.MessagePageResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/service.MessageResponse'
        type: array
      total:
        type: integer
    type: object
  service.MessageRequest:
    properties:
      content:
        type: string
      receiverId:
        type: integer
    required:
    - content
    - receiverId
    type: object
  service.MessageResponse:
    properties:
      content:
        type: string
      createdAt:
        type: string
      id:
        type: integer
      readed:
        type: boolean
      sender:
        $ref: '#/definitions/service.UserResponse'
        type: object
    type: object
...
  service.TagResponse:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  service.TopicNameResponse:
    properties:
      id:
        type: integer
      title:
        type: string
    type: object
...
  service.UserResponse:
    properties:
      avatar:
        type: string
      createdAt:
        type: string
      email:
        type: string
      fansNums:
        type: integer
      gender:
        type: integer
      id:
        type: integer
      introduce:
        type: string
      root:
        type: boolean
      username:
        type: string
      words:
        type: integer
      workDay:
        type: integer
      worksNums:
        type: integer
    type: object
...
  service.WorkResponse:
    properties:
      category:
        $ref: '#/definitions/service.CategoryResponse'
        type: object
      chapterNums:
        type: integer
      collegeNums:
        type: integer
      createdAt:
        type: string
      draftNums:
        type: integer
      id:
        type: integer
      introduce:
        type: string
      recycleNums:
        type: integer
      subscribeNums:
        type: integer
      subsectionNums:
        type: integer
      tags:
        items:
          $ref: '#/definitions/service.TagResponse'
        type: array
      talkNums:
        type: integer
      title:
        type: string
      topic:
        $ref: '#/definitions/service.TopicNameResponse'
        type: object
      type:
        type: integer
      updatedAt:
        type: string
      user:
        $ref: '#/definitions/service.UserResponse'
        type: object
      viewNums:
        type: integer
      words:
        type: integer
    type: object
info:
  contact: {}
  description: 属于国内的同人圈
  license: {}
  termsOfService: https://github.com/shyptr/archiveofourown
  title: 同人圈
  version: "1.0"
paths:
...
  /api/v1/college:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 页大小
        in: query
        name: pageSize
        type: integer
      - description: 当前页
        in: query
        name: page
        type: integer
      - description: 用户ID
        in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.CollegePageResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 查询书单列表
      tags:
      - 书单
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.CollegeCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 创建书单
      tags:
      - 书单
  /api/v1/college/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 书单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除
      tags:
      - 书单
    put:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 书单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.CollegeUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 修改书单信息
      tags:
      - 书单
  /api/v1/college/{id}/add/{workId}:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 书单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 作品ID
        in: path
        name: workId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 添加作品
      tags:
      - Chapter
  /api/v1/college/{id}/works:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 书单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        type: integer
      - description: 当前页
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.CollegeWorksPageResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取书单内作品列表
      tags:
      - 书单
  /api/v1/comment:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.CommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 添加评论
      tags:
      - 评论
  /api/v1/comment/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 评论ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 删除评论
      tags:
      - 评论
  /api/v1/comment/{objType}/{objID}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 评论对象类型,0-作品,1-章节,2-选段,3-公告
        enum:
        - 0
        - 1
        - 2
        - 3
        in: path
        name: objType
        required: true
        type: integer
      - description: 评论对象ID
        in: path
        name: objID
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 当前页
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.CommentPageResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 获取评论列表
      tags:
      - 评论
...
  /api/v1/message:
    post:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 参数
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/service.MessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: string
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 发送私信
      tags:
      - 私信
  /api/v1/message/{type}:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      - description: 私信类型
        enum:
        - 0
        - 1
        in: path
        name: type
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 当前页
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            $ref: '#/definitions/service.MessagePageResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 查看私信
      tags:
      - 私信
  /api/v1/message/unread:
    get:
      consumes:
      - application/json
      parameters:
      - description: token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功
          schema:
            type: int
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/errcode.Error'
        "500":
          description: 内部错误
          schema:
            $ref: '#/definitions/errcode.Error'
      summary: 未读私信数
      tags:
      - 私信
...
swagger: "2.0"